Done 1. Two Sum.
Takes in an input array of integers, as well as a target integer.
Attempts to find two indices such that they add up to the target
integer. If none is found, we return {0, 0}, as duplicates are not
allowed this will be immediately noticeable. The hints in the
question helped me realize that a Map was the way to go.