2215 difference of arrays done with time comparison
I was curious to see the fastest way to remove duplicates from an array,
with the additional caveat that the result has to be a vector. I found
that the fastest way on my computer was by making a constructor call to
an unordered set of the vector, then converting it back into a vector
and finally sorting it using std::sort. According to Leetcode, this is
much slower than doing a sort of the vector as well as erasing the
elements that are duplicates. Went for the latter solution.

Solution working and passing all tests, not represented in main of this
commit.