1207 Unique occurrences done
This one was to count the occurrences of integers for a given array, and
return whether that occurrence count was unique or not. I found this one
fairly trivial by using maps and sets, 0 ms run time and 11.35 MB
memory usage, which seems like this is a good way to solve the problem.