2095 delete middle of a linked list done
I did a naive approach first with a counter and looping over the entire
linked list to find the number of elements, to then know the middle. I
then googled and found a solution called the "Turtle and Hare"
algorithm. This algorithm simply has two pointers, where one moves at
twice the speed, and when that one reaches the end, the slow one is in
the middle. A third pointer was needed as the list only has access to
its next element and not the previous one.