392 - Is subsequence done
Tried to do this with both pointer arithmetic and indices. I got the
same 0 ms runtime in both examples. However indices said it used a lot
less memory, therefore it is an indices solution that is uploaded.

The original code does not use references to strings, which I find odd.
I also added const to signatures, as we are not looking to change either
of the strings sent in to the function.