1431 - greatest candies done
An easy problem where you have to find the largest element, and then
compare if the ith element plus a constant extraCandies is larger than
the largest element. I opted for the std function max rather than a for
loop, just to get some familiary with the functions in the standard
library. The alternative is a simple for loop, so I feel it does not go
against the spirit of LeetCode problems.