1456 - maximum vowels of substr k done
This one was interesting. I took the hint too literally at first and
made a vector that updated with each new iteration of the loop. I then
counted the vowels, leading to a real slow solution compared to everyone
else's. I then updated it to be a sliding window, where the count of the
previous symbol gets removed, and the next symbol gets added.